product page complete and working but with error that i can't get rid of
